Code quality: Perfect keyboard and mouse operation
